Use the LTS version of the SDK instead of the absolute latest (#433)
* Use the LTS version of the SDK instead of the absolute latest

* Add getLatestLTS function

* Rename to getLatestLTS

* Use npm instead of yarn -- issues in build

* Remove lockfiles

* export downloadedFrameworks
