+++ docs/ios/sidebar.js
@@ -19,12 +19,12 @@ module.exports = common({
     items: [{
       type: "link",
       label: "Swift", 
-      href: downloadedChangelog.cocoa.DittoSwift.latest.api_reference_docs_url,
+      href: downloadedChangelog.cocoa.DittoSwift["3.0.1"].api_reference_docs_url,
     },
     {
       type: "link",
       label: "ObjC", 
-      href: downloadedChangelog.cocoa.DittoObjC.latest.api_reference_docs_url,
+      href: downloadedChangelog.cocoa.DittoObjC["3.0.1"].api_reference_docs_url,
     }]
   },
 })
