+++ docs/raspberrypi/sidebar.js
@@ -5,6 +5,6 @@ module.exports = common({
   api: {
     type: "link",
     label: "API Reference", 
-    href: downloadedChangelog.cpp.Ditto.latest.api_reference_docs_url,
+    href: downloadedChangelog.cpp.Ditto["3.0.1"].api_reference_docs_url,
   }
 })
